move runtime exception to its own folder
