+++ core/common/src/main/java/alluxio/exception/runtime/UnauthenticatedRuntimeException.java
@@ -9,20 +9,18 @@
  * See the NOTICE file distributed with this work for information regarding copyright ownership.
  */
 
-package alluxio.exception.status;
+package alluxio.exception.runtime;
 
-import alluxio.exception.AlluxioRuntimeException;
 import alluxio.grpc.ErrorType;
 
 import io.grpc.Status;
 
 /**
- * Exception indicating that a client specified an invalid argument. Note that this differs from
- * FailedPreconditionException. It indicates arguments that are problematic regardless of the state
- * of the system (e.g., a malformed file name).
+ * Exception indicating that the request does not have valid authentication credentials for the
+ * operation.
  */
-public class InvalidArgumentRuntimeException extends AlluxioRuntimeException {
-  private static final Status STATUS = Status.INVALID_ARGUMENT;
+public class UnauthenticatedRuntimeException extends AlluxioRuntimeException {
+  private static final Status STATUS = Status.UNAUTHENTICATED;
   private static final ErrorType ERROR_TYPE = ErrorType.User;
   private static final boolean RETRYABLE = false;
 
@@ -30,7 +28,7 @@ public class InvalidArgumentRuntimeException extends AlluxioRuntimeException {
    * Constructor.
    * @param t cause
    */
-  public InvalidArgumentRuntimeException(Throwable t) {
+  public UnauthenticatedRuntimeException(Throwable t) {
     super(STATUS, null, t, ERROR_TYPE, RETRYABLE);
   }
 }
