--- core/common/src/main/java/alluxio/exception/status/InternalRuntimeException.java
@@ -1,52 +1,0 @@
-/*
- * The Alluxio Open Foundation licenses this work under the Apache License, version 2.0
- * (the "License"). You may not use this work except in compliance with the License, which is
- * available at www.apache.org/licenses/LICENSE-2.0
- *
- * This software is distributed on an "AS IS" basis, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND,
- * either express or implied, as more fully set forth in the License.
- *
- * See the NOTICE file distributed with this work for information regarding copyright ownership.
- */
-
-package alluxio.exception.status;
-
-import alluxio.exception.AlluxioRuntimeException;
-import alluxio.grpc.ErrorType;
-
-import io.grpc.Status;
-
-/**
- * Exception representing an internal error. This means some invariant expected by the underlying
- * system has been broken. If you see one of these errors, something is very broken.
- */
-public class InternalRuntimeException extends AlluxioRuntimeException {
-  private static final Status STATUS = Status.INTERNAL;
-  private static final ErrorType ERROR_TYPE = ErrorType.Internal;
-  private static final boolean RETRYABLE = false;
-
-  /**
-   * Constructor.
-   * @param message error message
-   * @param t cause
-   */
-  public InternalRuntimeException(String message, Throwable t) {
-    super(STATUS, message, t, ERROR_TYPE, RETRYABLE);
-  }
-
-  /**
-   * Constructor.
-   * @param message error message
-   */
-  public InternalRuntimeException(String message) {
-    super(STATUS, message, ERROR_TYPE, RETRYABLE);
-  }
-
-    /**
-   * Constructor.
-   * @param t cause
-   */
-  public InternalRuntimeException(Throwable t) {
-    super(STATUS, null, t, ERROR_TYPE, RETRYABLE);
-  }
-}
