+++ core/common/src/test/java/alluxio/worker/block/io/MockBlockWriter.java
@@ -11,7 +11,7 @@
 
 package alluxio.worker.block.io;
 
-import alluxio.exception.AlluxioRuntimeException;
+import alluxio.exception.runtime.AlluxioRuntimeException;
 import alluxio.network.protocol.databuffer.DataBuffer;
 
 import io.netty.buffer.ByteBuf;
