+++ core/server/proxy/src/main/java/alluxio/proxy/s3/S3ErrorResponse.java
@@ -11,7 +11,8 @@
 
 package alluxio.proxy.s3;
 
-import alluxio.exception.AlluxioRuntimeException;
+import alluxio.exception.runtime.AlluxioRuntimeException;
+import alluxio.exception.runtime.NotFoundRuntimeException;
 import alluxio.exception.status.AlluxioStatusException;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
@@ -152,7 +153,7 @@ public class S3ErrorResponse {
     XmlMapper mapper = new XmlMapper();
     S3ErrorCode s3ErrorCode;
     // TODO(WYY): we need to handle more exception in the future.
-    if (e instanceof alluxio.exception.status.NotFoundRuntimeException) {
+    if (e instanceof NotFoundRuntimeException) {
       // 404
       s3ErrorCode = S3ErrorCode.NO_SUCH_KEY;
     } else {
