+++ underfs/s3a/src/main/java/alluxio/underfs/s3a/AlluxioS3Exception.java
@@ -11,7 +11,7 @@
 
 package alluxio.underfs.s3a;
 
-import alluxio.exception.AlluxioRuntimeException;
+import alluxio.exception.runtime.AlluxioRuntimeException;
 import alluxio.grpc.ErrorType;
 
 import com.amazonaws.AmazonClientException;
